% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\docType{class}
\name{ClearScatterplot-class}
\alias{ClearScatterplot-class}
\title{ClearScatterplot Class}
\description{
A class for creating and managing scatterplot objects with enhanced visualization features.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing the data to be plotted.}

\item{\code{plot}}{An object storing the ggplot representation of the data.}
}}

