% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{create_plot}
\alias{create_plot}
\title{Create the Actual Plot}
\usage{
create_plot(
  data,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  logFoldChange = "log2fc",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)
}
\arguments{
\item{data}{The data frame containing the plot data.}

\item{color1}{Color for one category of data points.}

\item{color2}{Color for another category of data points.}

\item{color3}{Color for a third category of data points.}

\item{highLog2fc}{Threshold for high log2 fold change values.}

\item{lowLog2fc}{Threshold for low log2 fold change values.}

\item{expressionDirection}{Direction of gene expression.}

\item{negativeLogPValue}{The name of the column containing the negative log p-values.}

\item{logFoldChange}{The name of the column containing log fold change values.}

\item{timeVariable}{The variable representing time.}

\item{xAxis}{The name of the x-axis variable.}

\item{yAxis}{The name of the y-axis variable.}
}
\value{
The ggplot object.
}
\description{
This function creates the actual plot based on processed data and aesthetic choices.
}
