% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{createPlot,ClearScatterplot-method}
\alias{createPlot,ClearScatterplot-method}
\title{Create Plot Method for ClearScatterplot}
\usage{
\S4method{createPlot}{ClearScatterplot}(
  object,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)
}
\arguments{
\item{object}{A ClearScatterplot object.}

\item{color1}{Color for one category of data points.}

\item{color2}{Color for another category of data points.}

\item{color3}{Color for a third category of data points.}

\item{highLog2fc}{Threshold for high log2 fold change values.}

\item{lowLog2fc}{Threshold for low log2 fold change values.}

\item{negLog10pValue}{Threshold for -log10 p-value.}

\item{expressionDirection}{Direction of gene expression.}

\item{negativeLogPValue}{The name of the column containing the negative log p-values.}

\item{timeVariable}{The variable representing time.}

\item{xAxis}{The x-axis values.}

\item{yAxis}{The y-axis values.}
}
\value{
The ClearScatterplot object with the plot updated.
}
\description{
Creates a plot based on the ClearScatterplot object data.
}
\examples{
plotdata <- get_clear_scatterplot_df()
scatterplotObject <- ClearScatterplot(
  data = plotdata,
  logFoldChange = "log2fc",
  timePointColumn = "timePoint",
  timePointLevels = c("T10R1", "T5R1")
)
scattered_plot <- createPlot(
  scatterplotObject,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)
}
