% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{createPlot}
\alias{createPlot}
\title{Define a Generic Method 'createPlot'}
\usage{
createPlot(object, ...)
}
\arguments{
\item{object}{The object for which the plot is to be created. The class of this object
will determine which specific method is used.}

\item{...}{Additional arguments to be passed to the specific method for creating the plot.}
}
\value{
The result of the specific method for creating the plot, typically a plot object.
}
\description{
This generic function is designed to create plots for objects of various classes.
Specific methods should be defined for different classes to generate the appropriate plots.
}
