% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_clear_scatterplot_df}
\alias{get_clear_scatterplot_df}
\title{Provides a sample dataframe for clear scatterplot visualization}
\usage{
get_clear_scatterplot_df()
}
\value{
A dataframe with sample data for a clear scatterplot.
}
\description{
This function returns a predefined dataframe that represents data suitable for
creating a clear scatterplot. The dataframe includes log2 fold change values, p-values,
and regulation direction for a set of genes across different tissues and time points.
}
\examples{
clear_scatterplot_df <- get_clear_scatterplot_df()
}
