% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_informative_heatmap_df}
\alias{get_informative_heatmap_df}
\title{Provides a sample dataframe for informative heatmap visualization}
\usage{
get_informative_heatmap_df()
}
\value{
A dataframe with sample data for an informative heatmap.
}
\description{
This function returns a predefined dataframe that represents data suitable for
creating an informative heatmap. The dataframe includes values for three groups
and their associated p-values, with row names indicating gene identifiers.
}
\examples{
informative_heatmap_df <- get_informative_heatmap_df()
}
