% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_multifeature_grid_df}
\alias{get_multifeature_grid_df}
\title{Provides a sample dataframe for multifeature grid visualization}
\usage{
get_multifeature_grid_df()
}
\value{
A dataframe with sample data for a multifeature grid visualization.
}
\description{
This function returns a predefined dataframe that represents data suitable for
creating a multifeature grid visualization. The dataframe includes signaling pathways,
activation z-scores, and associated p-values for different tissues and time points.
}
\examples{
multifeature_grid_df <- get_multifeature_grid_df()
}
