% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformativeHeatmap.R
\name{getHeatmapObject,InformativeHeatmap-method}
\alias{getHeatmapObject,InformativeHeatmap-method}
\title{Retrieve the Heatmap Object from an InformativeHeatmap}
\usage{
\S4method{getHeatmapObject}{InformativeHeatmap}(x)
}
\arguments{
\item{x}{An \code{InformativeHeatmap} object from which the Heatmap object is to
be retrieved.}
}
\value{
A Heatmap object from the \code{ComplexHeatmap} package.
}
\description{
This method extracts the underlying Heatmap object stored within an
\code{InformativeHeatmap} object. It allows direct access to the \code{Heatmap} object
for further manipulation or inspection using \code{ComplexHeatmap} package
functionalities. Note that the \code{ComplexHeatmap} package is required
to fully utilize the returned Heatmap object.
}
\examples{
\donttest{
  # Assume `ih` is an existing InformativeHeatmap object
  heatmap_obj <- getHeatmapObject(ih)
  # Now `heatmap_obj` can be used directly with ComplexHeatmap functions
}

}
