% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformativeHeatmap.R
\name{getHeatmapObject}
\alias{getHeatmapObject}
\title{Define a Generic Method 'getHeatmapObject'}
\usage{
getHeatmapObject(x)
}
\arguments{
\item{x}{The object from which the Heatmap object is to be retrieved.
The class of this object will determine which specific method is used.}
}
\value{
The result of the specific method for retrieving the Heatmap object,
typically a Heatmap object.
}
\description{
This generic function is designed to retrieve the Heatmap object from
InformativeHeatmap objects. Specific methods should be defined for
different classes to extract the Heatmap object as needed.
}
