% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformativeHeatmap.R
\docType{class}
\name{InformativeHeatmap}
\alias{InformativeHeatmap}
\alias{InformativeHeatmap,ANY-method}
\title{InformativeHeatmap: A Class for Enhanced Heatmaps}
\usage{
\S4method{InformativeHeatmap}{ANY}(
  data,
  pch_val = 16,
  unit_val = 1,
  significant_color = "black",
  trending_color = "yellow",
  significant_pvalue = 0.05,
  trending_pvalue = 0.1,
  ...
)
}
\arguments{
\item{data}{A numeric matrix or data frame to be visualized as a heatmap.}

\item{pch_val}{Plotting character (pch) value for points in the heatmap.
Default is 16.}

\item{unit_val}{Size of the points in the heatmap. Specified in 'mm'.
Default is 1.}

\item{significant_color}{Color to be used for points representing significant values.
Default is "black".}

\item{trending_color}{Color to be used for points representing trending values
(significant but less so than those meeting the \code{significant_pvalue} criterion).
Default is "yellow".}

\item{significant_pvalue}{P-value threshold for significance.
Points with p-values below this threshold will be colored with \code{significant_color}.
Default is 0.05.}

\item{trending_pvalue}{P-value threshold for trending significance.
Points with p-values above \code{significant_pvalue} and below this threshold
will be colored with \code{trending_color}. Default is 0.1.}

\item{...}{Additional parameters passed to \code{ComplexHeatmap::Heatmap}.}
}
\value{
An object of class \code{InformativeHeatmap} with the heatmap initialized
and customized according to the provided parameters.
}
\description{
This class encapsulates the ComplexHeatmap package's Heatmap object,
allowing for extended functionality and customizations.

Constructor for the \code{InformativeHeatmap} class. This function initializes
an \code{InformativeHeatmap} object with the given data and additional parameters
for heatmap customization, including visual customization based on significance levels.
}
\section{Slots}{

\describe{
\item{\code{heatmap}}{Heatmap object from ComplexHeatmap package.}

\item{\code{params}}{List of parameters used to construct the heatmap.}
}}

\examples{
\donttest{
  data <- matrix(rnorm(100), ncol = 10)
  heatmap <- InformativeHeatmap(data, pch_val = 20, unit_val = 2,
                                significant_color = "red",
                                trending_color = "blue",
                                significant_pvalue = 0.05,
                                trending_pvalue = 0.1)
}
}
