% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedHeatmap.R
\name{AnnotatedHeatmapFromMAE}
\alias{AnnotatedHeatmapFromMAE}
\title{Create an AnnotatedHeatmap from a MultiAssayExperiment Object}
\usage{
AnnotatedHeatmapFromMAE(
  mae,
  significant_pvalue = 0.05,
  trending_pvalue = 0.1,
  significant_color = "black",
  trending_color = "yellow",
  pch_val = 16,
  unit_val = 4,
  K = 3,
  lambda = 0.2,
  coef = 2,
  ...
)
}
\arguments{
\item{mae}{A \code{MultiAssayExperiment} object containing multiple assays.}

\item{significant_pvalue}{P-value threshold for significance.
Points with p-values below this threshold will be colored with \code{significant_color}.
Default is 0.05.}

\item{trending_pvalue}{P-value threshold for trending significance.
Points with p-values above \code{significant_pvalue} and below this threshold
will be colored with \code{trending_color}. Default is 0.1.}

\item{significant_color}{Color to be used for points representing significant values.
Default is "black".}

\item{trending_color}{Color to be used for points representing trending values.
Default is "yellow".}

\item{pch_val}{Plotting character (pch) value for points in the heatmap. Default is 16.}

\item{unit_val}{Size of the points in the heatmap. Specified in 'mm'. Default is 1.}

\item{K}{Number of clusters for \code{iClusterPlus}. Default is 3.}

\item{lambda}{Regularization parameter for \code{iClusterPlus}. Default is 0.2.}

\item{coef}{Compare cluster group to be used as a contrast for \code{limma} analysis.}

\item{...}{Additional parameters passed to \code{ComplexHeatmap::Heatmap}.}
}
\value{
An object of class \code{AnnotatedHeatmap}.
}
\description{
This method takes a \code{MultiAssayExperiment} object, clusters the samples using \code{iClusterPlus},
compares the clusters using \code{limma}, and creates an \code{AnnotatedHeatmap} with a custom \code{layer_fun}
based on significance levels from the cluster comparison.
}
