% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultifeatureGrid.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Define a Generic Method 'plot_heatmap'}
\usage{
plot_heatmap(object, ...)
}
\arguments{
\item{object}{The object for which the plot is to be created. The class of this object
will determine which specific method is used.}

\item{...}{Additional arguments to be passed to the specific method for creating the plot.}
}
\value{
The result of the specific method for creating the plot, typically a ggplot object.
}
\description{
This generic function is designed for plotting the heatmap for MultifeatureGrid objects
}
