% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{show,ClearScatterplot-method}
\alias{show,ClearScatterplot-method}
\title{Display Method for ClearScatterplot}
\usage{
\S4method{show}{ClearScatterplot}(object)
}
\arguments{
\item{object}{A ClearScatterplot object.}
}
\value{
Prints the ggplot object stored in the plot slot of the ClearScatterplot.
The function itself returns invisible \code{NULL} to avoid additional console output.
}
\description{
This method displays the scatterplot stored in the ClearScatterplot object.
It invokes the \code{print} method on the \code{plot} object, which must be a ggplot object.
}
\examples{
# Create example data
plotdata <- get_clear_scatterplot_df()
p <- create_plot(
  plotdata,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  expressionDirection = "regulation",
  negativeLogPValue = "negLog10p",
  logFoldChange = "log2fc",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint"
)

# Display the plot
show(p)
}
