% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\docType{class}
\name{ClearScatterplot}
\alias{ClearScatterplot}
\title{ClearScatterplot: An S4 class for scatterplot visualization}
\description{
This class represents a scatterplot with specific data and plot configurations.
}
\section{Slots}{

\describe{
\item{\code{data_filepath}}{Character vector: Path to data file.}

\item{\code{data}}{Data.frame: data for the scatterplot.}

\item{\code{plot}}{ANY: The ggplot2 plot object (initially NULL).}
}}

