% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{ClearScatterplot}
\alias{ClearScatterplot}
\title{ClearScatterplot Constructor}
\usage{
ClearScatterplot(
  data,
  pValueColumn = "p",
  expressionColumnName = "log2fc",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  timePointColumn = "timePoint",
  timePointLevels = NULL,
  colorHigh = "cornflowerblue",
  colorNeutral = "grey",
  colorLow = "indianred",
  ...
)
}
\arguments{
\item{data}{A data frame containing the plot data.}

\item{pValueColumn}{The name of the column containing p-values.}

\item{expressionColumnName}{The name of the column containing expression values.}

\item{highLog2fc}{Threshold for high log2 fold change values.}

\item{lowLog2fc}{Threshold for low log2 fold change values.}

\item{negLog10pValue}{Threshold for -log10 p-value.}

\item{timePointColumn}{The name of the column containing time point information.}

\item{timePointLevels}{The levels for the time point column, if any.}

\item{colorHigh}{The color for high values.}

\item{colorNeutral}{The color for neutral values.}

\item{colorLow}{The color for low values.}
}
\value{
An object of class ClearScatterplot.
}
\description{
Creates an instance of the ClearScatterplot class.
}
\examples{
data <- data.frame(timePoint = c("T0", "T1"), p = runif(10), log2fc = runif(10, -2, 2))
scatterplot <- ClearScatterplot(data = data, pValueColumn = "p", expressionColumnName = "log2fc")
}
