% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{createPlot,ClearScatterplot-method}
\alias{createPlot,ClearScatterplot-method}
\title{Create and Assign a Plot to a ClearScatterplot Object}
\usage{
\S4method{createPlot}{ClearScatterplot}(
  object,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  expressionDirection = "regulation",
  timeVariable = "reg_time_org",
  xAxis = "organ",
  yAxis = "timePoint",
  changeMagnitude = "log2fc"
)
}
\arguments{
\item{object}{A \code{ClearScatterplot} object for which the plot will be created.}

\item{color1}{Character string specifying the color for high significance points. Default is "cornflowerblue".}

\item{color2}{Character string specifying the color for points that do not meet the high or low significance thresholds. Default is "grey".}

\item{color3}{Character string specifying the color for low significance points. Default is "indianred".}

\item{highLog2fc}{Numeric threshold for considering an expression value significantly high. Default is 0.585.}

\item{lowLog2fc}{Numeric threshold for considering an expression value significantly low. Default is -0.585.}

\item{expressionDirection}{Character string indicating the direction of regulation to be highlighted. Possible values are "up", "down", or "regulation" for both. Default is "regulation".}

\item{timeVariable}{Character string specifying the column in the data that represents the time variable. This parameter can be used to facet the plot by time points. Default is "reg_time_org".}

\item{xAxis}{Character string specifying the column name to be used as the x-axis variable. Default is "organ".}

\item{yAxis}{Character string specifying the column name to be used as the y-axis variable. Default is "timePoint".}

\item{changeMagnitude}{Character string specifying the column name to be used for the magnitude of change. This parameter allows dynamic specification of the x-axis variable in the plot. Default is "log2fc".}
}
\value{
The \code{ClearScatterplot} object with its \code{plot} slot updated to include the newly created plot.
}
\description{
Generates a scatter plot based on the stored data and aesthetic parameters provided. This plot is then assigned to the \code{plot} slot of the \code{ClearScatterplot} object.
}
\examples{
\dontrun{
  data_path <- system.file("extdata", "example_data.csv", package = "YourPackage")
  csPlot <- new("ClearScatterplot", data_filepath = data_path)
  csPlot <- createPlot(csPlot)
  # Now, csPlot contains a ggplot object in its plot slot that can be displayed using show(csPlot)
}
}
