% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{createPlot,ClearScatterplot-method}
\alias{createPlot,ClearScatterplot-method}
\title{Create Plot Method for ClearScatterplot}
\usage{
\S4method{createPlot}{ClearScatterplot}(
  object,
  color1 = "cornflowerblue",
  color2 = "grey",
  color3 = "indianred",
  highLog2fc = 0.585,
  lowLog2fc = -0.585,
  negLog10pValue = 1.301,
  expressionDirection = "regulation",
  timeVariable = "reg_time_org",
  xAxis = xAxis,
  yAxis = yAxis
)
}
\arguments{
\item{object}{A ClearScatterplot object.}

\item{color1}{Color for one category of data points.}

\item{color2}{Color for another category of data points.}

\item{color3}{Color for a third category of data points.}

\item{highLog2fc}{Threshold for high log2 fold change values.}

\item{lowLog2fc}{Threshold for low log2 fold change values.}

\item{expressionDirection}{Direction of gene expression.}

\item{timeVariable}{The variable representing time.}
}
\value{
The ClearScatterplot object with the plot updated.
}
\description{
Creates a plot based on the ClearScatterplot object data.
}
