% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedHeatmap.R
\name{getHeatmapObject}
\alias{getHeatmapObject}
\alias{getHeatmapObject,AnnotatedHeatmap-method}
\title{Define a Generic Method 'getHeatmapObject'}
\usage{
getHeatmapObject(x)

\S4method{getHeatmapObject}{AnnotatedHeatmap}(x)
}
\arguments{
\item{x}{An \code{AnnotatedHeatmap} object from which the Heatmap object is to
be retrieved.}
}
\value{
The result of the specific method for retrieving the Heatmap object,
typically a Heatmap object.

A Heatmap object from the \code{ComplexHeatmap} package.
}
\description{
This generic function is designed to retrieve the Heatmap object from
AnnotatedHeatmap objects. Specific methods should be defined for
different classes to extract the Heatmap object as needed.

This method extracts the underlying Heatmap object stored within an
\code{AnnotatedHeatmap} object. It allows direct access to the \code{Heatmap} object
for further manipulation or inspection using \code{ComplexHeatmap} package
functionalities. Note that the \code{ComplexHeatmap} package is required
to fully utilize the returned Heatmap object.
}
\examples{
  data <- matrix(rnorm(100), ncol = 10)
  ih <- AnnotatedHeatmap(data, pch_val = 20, unit_val = 2,
                                significant_color = "red",
                                trending_color = "blue",
                                significant_pvalue = 0.05,
                                trending_pvalue = 0.1)
  # Assume `ih` is an existing AnnotatedHeatmap object
  heatmap_obj <- getHeatmapObject(ih)
  # Now `heatmap_obj` can be used directly with ComplexHeatmap functions

}
