% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultifeatureGrid.R
\docType{class}
\name{MultifeatureGrid}
\alias{MultifeatureGrid}
\title{MultifeatureGrid: A Class for 2D Heatmap Visualization}
\usage{
MultifeatureGrid(
  data,
  title = "Heatmap",
  x_label = "X Label",
  y_label = "Y Label",
  logpval_label = "-log10(p-value)",
  zscore_label = "Activation z-score",
  numitems_label = "Number of Genes",
  color_palette = "RdYlBu",
  breaks = seq(-1, 1, 0.5)
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to be visualized in the heatmap.
The data frame should have columns that match the expected metadata fields
(e.g., significance values, z-scores).}

\item{title}{The title of the heatmap. Default is "Heatmap".}

\item{x_label}{The label for the x-axis. Default is "X Label".}

\item{y_label}{The label for the y-axis. Default is "Y Label".}

\item{logpval_label}{The label for the legend describing the -log10(p-value).
Default is "-log10(p-value)".}

\item{zscore_label}{The label for the legend describing the activation z-score.
Default is "Activation z-score".}

\item{numitems_label}{The label for the legend describing the number of genes
or items. Default is "Number of Genes".}

\item{color_palette}{The name of the color palette to use for the heatmap.
Default is "RdYlBu". This should be a valid palette name recognized by RColorBrewer.}

\item{breaks}{A numeric vector specifying the breakpoints for color mapping
across the heatmap's values. Default is \code{seq(-1, 1, 0.5)}.}
}
\value{
An object of class \code{MultifeatureGrid} initialized with the specified
parameters.
}
\description{
This class facilitates the creation of 2D heatmaps for visualizing biological data,
integrating significant levels and z-scores with customizable heatmap configurations.

This function initializes a \code{MultifeatureGrid} object with data and configuration
for creating a multi-feature grid heatmap. It allows customization of the heatmap's
title, axis labels, color palette, and more.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing heatmap data.}

\item{\code{title}}{The title of the heatmap.}

\item{\code{x_label}}{Label for the X-axis.}

\item{\code{y_label}}{Label for the Y-axis.}

\item{\code{logpval_label}}{Label for the legend of log10(p-value).}

\item{\code{zscore_label}}{Label for the legend of z-scores.}

\item{\code{numitems_label}}{Label for the legend describing the number of items (e.g., genes).}

\item{\code{color_palette}}{The name of the RColorBrewer color palette to be used.}

\item{\code{breaks}}{Numeric vector of breakpoints for color mapping.}
}}

\examples{
data <- data.frame(
  tissue = factor(rep(c("Tissue1", "Tissue2"), each = 4)),
  signaling = factor(rep(c("Pathway1", "Pathway2", "Pathway3", "Pathway4"), 2)),
  Activation_z_score = runif(8, -2, 2),
  p = runif(8, 0, 0.05),
  number_of_genes = sample(1:100, 8)
)
mg <- MultifeatureGrid(data)
plot_heatmap(mg)
\donttest{
  data <- data.frame(
    significance = runif(100),
    z_score = rnorm(100),
    num_genes = sample(1:100, 100, replace = TRUE)
  )
  mg <- MultifeatureGrid(data)
  # Further usage of mg...
}
}
