% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearScatterplot.R
\name{show,ClearScatterplot-method}
\alias{show,ClearScatterplot-method}
\title{Display the Scatterplot Stored in a ClearScatterplot Object}
\usage{
\S4method{show}{ClearScatterplot}(object)
}
\arguments{
\item{object}{A \code{ClearScatterplot} object containing the necessary data and configuration for generating the scatterplot.}
}
\value{
Invisible \code{ClearScatterplot} object. The primary purpose of this method is the side effect of displaying the plot, not returning a value. However, the object itself is returned invisibly for potential further manipulation or inspection.
}
\description{
This method invokes the plotting functionality to generate and display the scatterplot based on the data and configuration stored within the \code{ClearScatterplot} object. If the plot has not been created yet, it first calls the \code{createPlot} method to generate the plot and then displays it.
}
\examples{
\dontrun{
  data_path <- system.file("extdata", "example_data.csv", package = "YourPackage")
  csPlot <- new("ClearScatterplot", data_filepath = data_path)
  # Assuming createPlot method has been properly defined and exported
  show(csPlot)
  # This will display the scatterplot based on the data
}
}
