% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformativeHeatmap.R
\name{updateLayerFun}
\alias{updateLayerFun}
\title{Define a Generic Method 'updateLayerFun'}
\usage{
updateLayerFun(x, layer_fun)
}
\arguments{
\item{x}{An \code{InformativeHeatmap} object whose layer function is to be updated.}

\item{layer_fun}{A function that defines the new layer to be applied to the heatmap.
The class of this object will determine which specific method is used.}
}
\value{
Returns an updated \code{InformativeHeatmap} object with the new layer function applied.
}
\description{
This generic function is for updating the layer function of an existing
InformativeHeatmap objects.
}
